Made a cog for cats
